i12e: 'wgkey' code moved to new 'internal/mesh/wgkey' package
